Make a generic ThemedModal and more custom FloatingModal
