+++ codesign/components/ui/ThemedModal.tsx
@@ -1,14 +1,10 @@
-import { Modal, type ModalProps, StyleSheet, Pressable } from 'react-native';
+import { Modal, type ModalProps } from 'react-native';
 
-import { ThemedView } from '../ThemedView';
-import { Layout } from '@/constants/styles/Layout';
-import { Border } from '@/constants/styles/Border';
-import { Spacing } from '@/constants/styles/Spacing';
+import { useThemeColor } from '@/hooks/useThemeColor';
 
 export type ThemedModalProps = ModalProps & {
   lightColor?: string;
   darkColor?: string;
-  closeModal: () => void;
   children: React.ReactNode;
 };
 
@@ -16,28 +12,22 @@ export function ThemedModal({
   lightColor,
   darkColor,
   children,
-  closeModal,
+  style,
   ...otherProps
 }: ThemedModalProps) {
+  const backgroundColor = useThemeColor(
+    { light: lightColor, dark: darkColor },
+    'background'
+  );
+
   return (
-    <Modal transparent={true} animationType="fade" {...otherProps}>
-      <Pressable onPress={closeModal} style={styles.modalContainer}>
-        <ThemedView style={styles.modal}>{children}</ThemedView>
-      </Pressable>
+    <Modal
+      transparent={true}
+      animationType="fade"
+      {...otherProps}
+      style={[{ backgroundColor }, style]}
+    >
+      {children}
     </Modal>
   );
 }
-
-const styles = StyleSheet.create({
-  modalContainer: {
-    ...Layout.flex,
-    ...Layout.center,
-    backgroundColor: 'rgba(0, 0, 0, 0.3)'
-  },
-  modal: {
-    ...Layout.center,
-    ...Border.roundedSmall,
-    ...Border.elevated,
-    padding: Spacing.large
-  }
-});
